requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
albumentations==0.4.3
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
astropy==6.0.1
astropy-iers-data==0.2024.4.1.0.33.14
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
blobfile==2.1.1
certifi==2022.12.7
charset-normalizer==2.1.1
clean-fid==0.1.35
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
comm==0.2.2
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.8
-e git+https://github.com/alexzhou907/DDBM.git@0d3cc96b852323e17250a7b5e29cc9e42e078bfa#egg=ddbm
debugpy==1.8.0
decorator==5.1.1
deepinv==0.1.1
docker-pycreds==0.4.0
ehtim==1.2.7
ehtplot @ git+https://github.com/liamedeiros/ehtplot.git@5222b90e353e3e34d05746b7add1f3628ac945e0
einops==0.7.0
exceptiongroup==1.2.1
executing==2.0.1
filelock==3.9.0
flash-attn==2.0.4
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2023.4.0
ftfy==6.1.3
future==1.0.0
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.42
h5py==3.10.0
hdf5storage==0.1.19
hdrhistogram==0.10.3
huggingface-hub==0.20.3
hydra-core==1.3.2
idna==3.4
imageio==2.34.0
imgaug==0.2.6
iniconfig==2.0.0
iopath==0.1.10
ipykernel==6.29.4
ipython==8.23.0
jedi==0.19.1
Jinja2==3.1.2
joblib==0.14.0
jplephem==2.21
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
lazy_loader==0.3
lightning-utilities==0.10.1
lmdb==1.4.1
lxml==4.9.4
MarkupSafe==2.1.3
matplotlib==3.8.3
matplotlib-inline==0.1.7
monai==1.3.0
mpi4py @ file:///croot/mpi4py_1671223370575/work
mpmath==1.3.0
multidict==6.0.5
nest-asyncio==1.6.0
networkx==3.2.1
nfft==0.1
ninja==1.11.1.1
numpy==1.26.3
omegaconf==2.3.0
opencv-python-headless==4.9.0.80
packaging==23.2
pandas==1.5.3
pandas-appender==0.9.8.4
paramsurvey==0.4.19
parso==0.8.4
pbr==6.0.0
pexpect==4.9.0
pillow==10.2.0
piq==0.7.0
platformdirs==4.2.0
pluggy==1.5.0
portalocker==2.8.2
prdc==0.2
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycryptodomex==3.20.0
pyerfa==2.0.1.2
Pygments==2.17.2
pylops==2.2.0
pyNFFT @ file:///home/conda/feedstock_root/build_artifacts/pynfft_1650036921293/work
pyparsing==3.1.1
pytest==8.2.0
python-dateutil==2.8.2
pytorch-lightning==2.2.1
pytz==2024.1
PyWavelets==1.5.0
PyYAML==6.0.1
pyzmq==26.0.1
regex==2023.12.25
requests==2.28.1
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.0.2
scipy==1.12.0
sentry-sdk==1.40.6
setproctitle==1.3.3
sgp4==2.23
six==1.16.0
skyfield==1.48
smmap==5.0.1
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
taming-transformers==0.0.1
termcolor==2.4.0
threadpoolctl==3.3.0
tifffile==2024.2.12
tomli==2.0.1
torch==2.1.0+cu121
torch-fidelity==0.3.0
torchaudio==2.1.0+cu121
torchmetrics==1.3.1
torchvision==0.16.0+cu121
tornado==6.4
tqdm==4.66.1
traitlets==5.14.3
triton==2.1.0
typing_extensions==4.8.0
tzdata==2024.1
urllib3==1.26.13
wandb==0.16.3
wcwidth==0.2.13
yacs==0.1.8
yarl==1.9.4
